Interaction areas, Dialogue, NPC, Box updates
I added interactable areas so if the player presses "e" when inside of the area they can initiate an interaction. I also added an NPC which when interacted with will start a dialogue. Lastly I updated the box to turn semi-transparent when the player stands behind it, this way the player knows where they are at all times. I also updated the box text to display "hidden" only when the player is crouching.